added a min.c file

+++ lab6/min.c
@@ -6,7 +6,7 @@ int main(int argc,char *argv[])
     int i=1;int j=0;
     int m=0; 
     double a[100];
-    int temp;
+    double temp;
     for(i=1;i<argc;i++,j++){
     a[j]=atof(argv[i]);
     }
@@ -15,7 +15,7 @@ int main(int argc,char *argv[])
             temp=a[m]; a[m]=a[m+1];a[m+1]=temp;
         }
     }
-    printf("The min is%d\n",temp);
+    printf("The min is%g\n",temp);
     return 0;
 }
